requirements.txt
python-dotenv==1.0.1

qdrant-client==1.9.1
openai==1.30.1

langgraph==0.1.19
langchain==0.2.0
langchain-openai==0.1.7

fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

redis==5.0.4
numpy==1.26.4

asyncpg==0.29.0
psycopg2-binary==2.9.9
faker==24.0.0

bcrypt==4.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
